# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz API",
        "description": "API Documentation",
        "contact": {
            "name": "API Support",
            "email": "marko.mandic.engr@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/images": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get a paginated list of images",
                "description": "Returns a paginated list of all images",
                "operationId": "listImages",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Image"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 75
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Images"
                ],
                "summary": "Create a new image",
                "description": "Creates a new image and returns the created image",
                "operationId": "storeImage",
                "requestBody": {
                    "description": "Image data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreImageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Image created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Image created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Image"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/images/{id}": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get a specific image",
                "description": "Returns a specific image",
                "operationId": "getImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the image to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Image"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Image not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Images"
                ],
                "summary": "Update an image",
                "description": "Updates an image and returns the updated image",
                "operationId": "updateImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the image to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Image data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateImageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Image updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Image"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Image not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Images"
                ],
                "summary": "Delete an image",
                "description": "Deletes an image and returns a success message",
                "operationId": "deleteImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the image to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Image deleted successfully"
                    },
                    "404": {
                        "description": "Image not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreImageRequest": {
                "required": [
                    "url"
                ],
                "properties": {
                    "url": {
                        "description": "URL of the image",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/images/chair.jpg"
                    }
                },
                "type": "object"
            },
            "UpdateImageRequest": {
                "required": [
                    "url"
                ],
                "properties": {
                    "url": {
                        "description": "URL of the image",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/images/updated-chair.jpg"
                    }
                },
                "type": "object"
            },
...
            "Image": {
                "required": [
                    "id",
                    "url"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the image",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "url": {
                        "description": "URL of the image",
                        "type": "string",
                        "example": "https://example.com/images/chair.jpg"
                    },
                    "created_at": {
                        "description": "When the image was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "When the image was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}